backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
pymongo==4.6.2
python-dotenv==1.0.1
requests==2.31.0
email-validator==2.1.1
